+++ src/message_filters/__init__.py
@@ -39,10 +39,13 @@ import itertools
 import threading
 import rclpy
 
-import builtin_interfaces
+from typing import Type, Union
+
 from rclpy.clock import ROSClock
 from rclpy.duration import Duration
 from rclpy.logging import LoggingSeverity
+from rclpy.node import Node
+from rclpy.qos import QoSProfile
 from rclpy.time import Time
 
 
@@ -76,12 +79,34 @@ class Subscriber(SimpleFilter):
     from a ROS2 subscription through to the filters which have connected
     to it.
     """
-    def __init__(self, *args, **kwargs):
+
+    def __init__(
+        self,
+        node: Node,
+        msg_type: Type,
+        topic: str,
+        qos_profile: Union[QoSProfile, int] = QoSProfile(depth=10),
+    ) -> None:
+        """
+        Construct a Subscriber.
+
+        :param node: The node to create a subscriber for.
+        :param msg_type: The type of ROS messages the subscription will subscribe to.
+        :param topic: The name of the topic the subscription will subscribe to.
+        :param qos_profile: A QoSProfile or a history depth to apply to the
+            subscription. In the case that a history depth is provided, the QoS history is
+            set to KEEP_LAST, the QoS history depth is set to the value of the parameter,
+            and all other QoS settings are set to their default values.
+        """
         SimpleFilter.__init__(self)
-        self.node = args[0]
-        self.topic = args[2]
-        kwargs.setdefault('qos_profile', 10)
-        self.sub = self.node.create_subscription(*args[1:], self.callback, **kwargs)
+        self.node = node
+        self.topic = topic
+        self.sub = self.node.create_subscription(
+            msg_type=msg_type,
+            topic=self.topic,
+            callback=self.callback,
+            qos_profile=qos_profile,
+        )
 
     def callback(self, msg):
         self.signalMessage(msg)
@@ -217,10 +242,11 @@ class Chain(SimpleFilter):
 
     def __init__(self, message_filter=None):
         SimpleFilter.__init__(self)
-        if message_filter is not None:
-            self.connectInput(message_filter)
 
         self.incoming_connection = None
+
+        if message_filter is not None:
+            self.connectInput(message_filter)
 
         self._message_filters: dict[int, Chain.FilterInfo] = {}
 
@@ -297,12 +323,25 @@ class TimeSynchronizer(SimpleFilter):
             del my_queue[min(my_queue)]
         # common is the set of timestamps that occur in all queues
         common = reduce(set.intersection, [set(q) for q in self.queues])
+        signaled_time = None
         for t in sorted(common):
             # msgs is list of msgs (one from each queue) with stamp t
             msgs = [q[t] for q in self.queues]
             self.signalMessage(*msgs)
             for q in self.queues:
                 del q[t]
+            signaled_time = t
+            break
+
+        # for consistency with the C++ implementation:
+        #     Delete all stored messages with a timestamp
+        #     older than the time of the latest signaled time
+        if signaled_time is not None:
+            for queue in self.queues:
+                for stamp in list(queue.keys()):
+                    if stamp < signaled_time:
+                        del queue[stamp]
+
         self.lock.release()
 
 class ApproximateTimeSynchronizer(TimeSynchronizer):